fix: prevent concurrent modification exception
These fixes prevent concurrent modification
exceptions and allow for additional commands
to be scheduled while processing the
toSchedule command queue.

These fixes are not perfect, and a review
of the current scheduling handling is needed
in order to better handle throughput across
both the core scheduler, and sub scheduling
that is done in meta command constructs
themselves.